fix(build): explicit narrowing cast for TEB32.WOW32Reserved write
MSVC's /W4 /WX treated the implicit uint64_t->uint32_t conversion
(image_base + w64svc_turbo_bop_rva assigned into the 32-bit
WOW32Reserved field) as an error, breaking every Windows build target.
Clang/macOS didn't flag it, so this wasn't caught locally.

+++ src/windows-emulator/module/module_manager.cpp
@@ -821,7 +821,9 @@ namespace sogen
                 {
                     if (t.teb32.has_value())
                     {
-                        t.teb32->access([&](TEB32& teb32_obj) { teb32_obj.WOW32Reserved = image_base + w64svc_turbo_bop_rva; });
+                        t.teb32->access([&](TEB32& teb32_obj) {
+                            teb32_obj.WOW32Reserved = static_cast<std::uint32_t>(image_base + w64svc_turbo_bop_rva);
+                        });
                     }
                 }
